Streamlined the code a bit for cost distribution
