+++ src/pages/costModels/createCostModelWizard/review.tsx
@@ -51,7 +51,7 @@ const ReviewSuccess = withTranslation()(ReviewSuccessBase);
 
 const ReviewDetailsBase: React.SFC<WithTranslation> = ({ t }) => (
   <CostModelContext.Consumer>
-    {({ name, description, type, markup, sources, tiers, createError, isDiscount }) => {
+    {({ name, description, distribution, type, markup, sources, tiers, createError, isDiscount }) => {
       return (
         <>
           {createError && <Alert variant="danger" title={`${createError}`} />}
@@ -92,6 +92,14 @@ const ReviewDetailsBase: React.SFC<WithTranslation> = ({ t }) => (
                     {t('cost_models_wizard.review.markup')}
                   </TextListItem>
                   <TextListItem component={TextListItemVariants.dd}>{isDiscount ? '-' + markup : markup}%</TextListItem>
+                  {type === 'OCP' && (
+                    <>
+                      <TextListItem component={TextListItemVariants.dt}>
+                        {t('cost_models_details.distribution_type')}
+                      </TextListItem>
+                      <TextListItem component={TextListItemVariants.dd}>{distribution}</TextListItem>
+                    </>
+                  )}
                   <TextListItem component={TextListItemVariants.dt}>
                     {t('cost_models_wizard.review.sources')}{' '}
                     {sources.find(src => src.selected && Boolean(src.costmodel)) && (
@@ -116,7 +124,7 @@ const ReviewDetailsBase: React.SFC<WithTranslation> = ({ t }) => (
 
 const ReviewDetails = withTranslation()(ReviewDetailsBase);
 
-const Review = () => {
+const ReviewWithDistribution = () => {
   return (
     <CostModelContext.Consumer>
       {({ createSuccess }) => {
@@ -129,4 +137,4 @@ const Review = () => {
   );
 };
 
-export default Review;
+export default ReviewWithDistribution;
